add @types for jest
