+++ package.json
@@ -9,6 +9,7 @@
 	"devDependencies": {
 		"@types/classnames": "2.2.3",
 		"@types/date-fns": "2.6.0",
+		"@types/jest": "^22.1.3",
 		"@types/react": "16.0.34",
 		"@types/react-dom": "16.0.3",
 		"@types/react-redux": "5.0.14",
